% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s17}
\alias{s17}
\title{2017 NYC School Survey Results Joined with Census and School Meta Data}
\format{A data frame with 401 rows and 133 variables:
\describe{
  \item{open_year}{Year/month/day School was established}
  \item{years_open}{Number of years open (since 2018)}
  \item{doe_grade_levels}{grade levels offered in this school}
  
  \item{pct_any_health_care_u65}{Census: Percent of school Census tract with with public or private healthcare}
  \item{pct_unv_deg}{Census: Percent of school Census tract with a university degree}
  \item{pct_under_hs}{Census: Percent of school Census tract with less than HS education}
  \item{med_house_income}{Census: Estimated median household income of school Census tract}
  \item{pct_poverty}{Census: Percent of school Census tract below poverty line}
  \item{pct_income_100k_up}{Census: Percent of school Census tract with income 100k or higher}
  \item{pct_income_150k_up}{Census: Percent of school Census tract with income 150k or higher}
  \item{pct_black/white/amind/asian/nhpi/oth/latin/pct_mlt_r/nonwht}{Census: Percent of school Census tract within by Census race/rthnicity categories}
  
  \item{parent_resp_rate}{Parent survey response rate}
  \item{student_resp_rate}{Student survey response rate}
  \item{total_resp_rate}{Total survey response rate}
  
  \item{colab}{Composite score for the school: Collaborative Teachers}
  \item{eff_leaders}{Composite score for the school: Effective leadership}
  \item{rig_instr}{Composite score for the school: Rigorous Instruction}
  \item{supp_env}{Composite score for the school: Supportive Environment}
  \item{stg_ties}{Composite score for the school: Strong Family/Comminity Ties}
  \item{trust}{Composite score for the school: Trust Score}
  
  \item{p_q*_score}{Responses from the Parent survey on standardized 0-100 point scale}
  \item{s_q*_score}{Responses from the Student survey on standardized 0-100 point scale}
  \item{lat}{DOE Points File: Latitude of school}
  \item{lng}{DOE Points File: Longitude of school}
  \item{year}{Year of Survey}
  ...
}}
\source{
http://schools.nyc.gov/Accountability/tools/survey/2017Survey.htm
}
\usage{
s17
}
\description{
This is a subset of parent and student survey questions from the New York City
School Survey. Non-highschool institutions are dropped, and the remaining records
are joined with Census variables and information about each
participating school from a meta file provided by the DOE. See the git repository
for this project for additional source files and the data dictionary:
https://github.com/DerekYves/nycschools
}
\keyword{datasets}
