% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score_calc}
\alias{score_calc}
\title{Calculate percent positive response for a question in the New York City School Survey data}
\usage{
score_calc(df, stub)
}
\arguments{
\item{df}{data.frame; s16 or s17}

\item{stub}{character; The question stub of interest, e.g. "s_q5c", which is the stub for "5c. My teachers will always listen to students' ideas."}
}
\value{
A numeric vector of proportions with console output indicating the selected variables
}
\description{
Calculate percent positive response for a question in the New York City School Survey data
}
\examples{
\dontrun{
df <- s17
df$test <- score_calc(df,  stub="s_q5c")
}
}
